% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_f.R
\name{boot_f}
\alias{boot_f}
\title{boot_f}
\usage{
boot_f(data, indices, fun, cor.type)
}
\arguments{
\item{data}{The data as a vector, matrix or data frame. If it is a matrix or data frame then each
row is considered as one multivariate observation.}

\item{indices}{indices to be used for calculation}

\item{fun}{a function to use in bootstrap procedure
coefficient is to be computed. One of "pearson" (default), "kendall", or
"spearman".}

\item{cor.type}{a character string indicating which correlation}
}
\value{
A matrix with bootstraped estimates
}
\description{
Generate R bootstrap replicates of a statistic applied to data. Function is based on boot() from
boot R package
}
\references{
https://www.datacamp.com/community/tutorials/bootstrap-r
}
\keyword{internal}
