% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_f_brnn.R
\name{boot_f_brnn}
\alias{boot_f_brnn}
\title{boot_f_brnn}
\usage{
boot_f_brnn(data, index, brnn.formula, neurons = 3)
}
\arguments{
\item{data}{data frame with variables for model fitting}

\item{index}{indices to be used for calculation}

\item{brnn.formula}{object of class formula to be passed into brnn function}
}
\value{
A matrix with bootstrapped estimates of r squared and adjusted r squared
}
\description{
Generate R bootstrap replicates of a brnn function applied to data. Function is based on boot() from
boot R package
}
\references{
https://www.datacamp.com/community/tutorials/bootstrap-r
}
\keyword{internal}
