% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_f_cor.R
\name{boot_f_cor}
\alias{boot_f_cor}
\title{boot_f_cor}
\usage{
boot_f_cor(data, indices, cor.type)
}
\arguments{
\item{data}{The data as a vector, matrix or data frame. If it is a matrix or data frame then each
row is considered as one multivariate observation.}

\item{indices}{indices to be used for calculation}

\item{cor.type}{a character string indicating which correlation}

\item{fun}{a function to use in bootstrap procedure
coefficient is to be computed. One of "pearson" (default), "kendall", or
"spearman".}
}
\value{
A matrix with bootstrapped estimates of correlation coefficients
}
\description{
Generate R bootstrap replicates of a statistic applied to data.
}
\references{
https://www.datacamp.com/community/tutorials/bootstrap-r
}
\keyword{internal}
