% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_f_lm.R
\name{boot_f_lm}
\alias{boot_f_lm}
\title{boot_f_lm}
\usage{
boot_f_lm(data, index, lm.formula)
}
\arguments{
\item{data}{data frame with variables for model fitting}

\item{index}{indices to be used for calculation}

\item{lm.formula}{object of class formula to be passed into lm function}
}
\value{
A matrix with bootstrapped estimates of r squared and adjusted r squared
}
\description{
Generate R bootstrap replicates of a lm function applied to data.
}
\references{
https://www.datacamp.com/community/tutorials/bootstrap-r
}
\keyword{internal}
