% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_f_pcor.R
\name{boot_f_pcor}
\alias{boot_f_pcor}
\title{boot_f_pcor}
\usage{
boot_f_pcor(data, indices, cor.type)
}
\arguments{
\item{data}{A vector, matrix, or data frame containing the input data.
If a matrix or data frame is provided, each row is treated as a separate
multivariate observation.}

\item{indices}{A numeric or integer vector specifying the indices to be
utilized in the calculation. These indices determine which elements or rows
of the 'data' are to be included in the analysis}

\item{cor.type}{A string specifying the method of correlation to be applied
in the bootstrap procedure. Available options are "pearson" (default),
"kendall", or "spearman".}
}
\value{
A matrix with bootstrapped estimates of partial correlation coefficients
}
\description{
Generate R bootstrap replicates of a statistic applied to data.
}
\references{
https://www.datacamp.com/community/tutorials/bootstrap-r
}
\keyword{internal}
