% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ones.R
\name{count_ones}
\alias{count_ones}
\title{count_ones}
\usage{
count_ones(vector)
}
\arguments{
\item{vector}{a vector of intigers}
}
\value{
an intiger of counted ones
}
\description{
calculates share of intiger 1 in a vector
}
