% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.r
\name{critical_r}
\alias{critical_r}
\title{critical_r}
\usage{
critical_r(n, alpha = 0.05)
}
\arguments{
\item{n}{number of observations}

\item{alpha}{significance level}
}
\value{
calculated critical value of Pearson correlation coefficient
}
\description{
Calculates critical value of Pearson correlation coefficient
for a selected alpha.
}
\examples{
threshold_1 <- critical_r(n = 55, alpha = 0.01)
threshold_2 <- critical_r(n = 55, alpha = 0.05)
}
