% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MVA.R
\docType{data}
\name{data_MVA}
\alias{data_MVA}
\title{Mean vessel area example proxy from 2012 - 1940}
\format{
A data frame with 73 rows and 1 variable:
\describe{
  \item{MVA}{Mean vessel area [mm^2] indices from 2012 - 1940}
}
}
\source{
Jernej Jevšenak, Slovenian Forestry Institute, Večna pot 2,
Ljubljana, Slovenia
}
\usage{
data_MVA
}
\description{
A dataset with MVA proxy records from a lowland forest Mlače in Slovenia.
The first row represents a value of a year in 2012. Row names represent
years.
}
\keyword{datasets}
