% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_TRW_1.R
\docType{data}
\name{data_TRW_1}
\alias{data_TRW_1}
\title{Tree-ring width (TRW) data from 2012 - 1961}
\format{
A data frame with 52 rows and 1 variable:
\describe{
  \item{TRW}{Standardized tree-ring width indices from 2012 - 1961}
}
}
\source{
Tom Levanič, Slovenian Forestry Institute, Večna pot 2,
Ljubljana, Slovenia
}
\usage{
data_TRW_1
}
\description{
A dataset of tree-ring widths (TRW) from a site in Krakovo forest (Slovenia).
The first row represents a value of a year in 1961.
}
\keyword{datasets}
