% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_MVA.R
\docType{data}
\name{dataset_MVA}
\alias{dataset_MVA}
\title{MVA and mean April temperature}
\format{
A data frame with 79 rows and 2 variables:
\describe{
  \item{MVA}{Mean vessel area measurements from 2012 - 1934}
  \item{T_Apr}{Mean April temperature for the meteorological station Maribor from 2012 - 1934}
}
}
\source{
Jevšenak J., Levanič T. 2015. Dendrochronological and wood-anatomical features of
differently vital pedunculate oak (Quercus robur L.) stands and their response to climate.
Topola, 195/196: 85-96
}
\usage{
dataset_MVA
}
\description{
A dataset with a mean vessel area (MVA) chronology of Quercus robur from
a lowland oak forest in Eastern Slovenia and a mean April temperature.
This dataset includes years for the period 2012-1934. For a detailed
description about the MVA chronology development, sampling site and the
calculations of mean monthly correlations, see Jevšenak and Levanič (2015).
}
\keyword{datasets}
