% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_TRW.R
\docType{data}
\name{dataset_TRW}
\alias{dataset_TRW}
\title{TRW and mean June -  July temperature from Albania}
\format{A data frame with 59 rows and 2 variables:
\describe{
  \item{TRW}{Standardised tree-ring width chronology of Pinus nigra from Albania}
  \item{T_Jun_Jul}{Mean June - July temperature for Albania downloaded from KNMI Climate Explorer}
}}
\source{
Levanič, T., Poljanšek, S., Toromani, E., 2015. Early summer temperatures reconstructed from
black pine (Pinus nigra Arnold) tree-ring widths from Albania. The Holocene 25, 469-481.
}
\usage{
dataset_TRW
}
\description{
A dataset with a tree-ring width (TRW) chronology of Pinus nigra from Albania and mean
June-July temperature. This TRW chronology has a span of 59 years (period 2009 - 1951)
and was already used to reconstruct summer temperatures by Levanič et al. (2015).
In this paper, all the details about sample replication, site description and correlation
statistics are described.
}
\keyword{datasets}
