% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_TRW_complete.R
\docType{data}
\name{dataset_TRW_complete}
\alias{dataset_TRW_complete}
\title{The complete dataset of standardized tree-ring chronology from Albania}
\format{A data frame with 551 rows and 1 variable:
\describe{
  \item{TRW}{Standardised tree-ring width chronology of Pinus nigra from Albania}
}}
\source{
Levanič, T., Poljanšek, S., Toromani, E., 2015. Early summer temperatures reconstructed from
black pine (Pinus nigra Arnold) tree-ring widths from Albania. The Holocene 25, 469-481.
}
\usage{
dataset_TRW_complete
}
\description{
A dataset with a tree-ring width (TRW) chronology of Pinus nigra from Albania This TRW
chronology has a span of 551 years (period 2009 - 1459) and was already used to reconstruct
summer temperatures by Levanič et al. (2015). In this paper, all the details about sample
replication, site description and correlation statistics are described.
}
\keyword{datasets}
