% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset_1.R
\docType{data}
\name{example_dataset_1}
\alias{example_dataset_1}
\title{Example of dataset as required for compare_methods()}
\format{A data frame with 58 rows and 3 columns :
\describe{
  \item{MVA}{Mean Vessel Area measurements from 2012 - 1955}
  \item{T_APR}{Mean April temperatures from 2012 - 1955}
  \item{T_aug_sep}{Mean August-September temperatures from preceding
  growing season from 2012 - 1955}
}}
\source{
Jernej Jevšenak, Slovenian Forestry Institute, Večna pot 2,
Ljubljana, Slovenia
}
\usage{
example_dataset_1
}
\description{
A dataset of Mean Vessel Area (MVA) tree-ring parameter from a lowland
forest in Slovenia. The first row represents a value of a year in 2012.
}
\keyword{datasets}
