% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_MVA.R
\docType{data}
\name{example_MVA}
\alias{example_MVA}
\title{Mean vessel area example proxy from 2012 - 1961}
\format{A data frame with 52 rows and 1 variable:
\describe{
  \item{MVA}{Mean vessel area [mm^2] indices from 2012 - 1961}
}}
\source{
prof. dr. Tom Levanic, Slovenian Forestry Institute, Vecna pot 2,
Ljubljana, Slovenia
}
\usage{
example_MVA
}
\description{
A dataset with MVA proxy records from a lowland forest Mlace in Slovenia.
The first row represents a value of a year in 1961. Row names represent
years.
}
\keyword{datasets}
