% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_proxies_1.R
\docType{data}
\name{example_proxies_1}
\alias{example_proxies_1}
\title{Tree-ring example proxies 1 from 2015 - 1961}
\format{A data frame with 55 rows and 3 variables:
\describe{
  \item{MVA}{Mean vessel area [mm^2] indices from 2015 - 1961}
  \item{O18}{Scaled Stable oxygen isotope ratios from 2015 - 1961}
  \item{TRW}{Tree-ring widths from 2015 - 1961}
}}
\source{
Jernej Jevšenak, Slovenian Forestry Institute, Večna pot 2,
Ljubljana, Slovenia
}
\usage{
example_proxies_1
}
\description{
A dataset with three tree-ring proxy records from a site near Ljubljana
(Slovenia). The first row represents a value of a year in 1961. The three
proxy records are MVA (Mean vessel area [mm ^2]), O (stable oxygen isotope
ratios) and TRW (Tree-ring widths)
}
\keyword{datasets}
