% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_proxies_individual.R
\docType{data}
\name{example_proxies_individual}
\alias{example_proxies_individual}
\title{Example of dataset with individual chronologies of MVA.}
\format{A data frame with 56 rows and 54 columns :
\describe{
  \item{MVA_1}{Mean vessel area chronology for tree 1}
  \item{MVA_2}{Mean vessel area chronology for tree 2}
  \item{MVA_3}{Mean vessel area chronology for tree 3}
  \item{MVA_4}{Mean vessel area chronology for tree 4}
  \item{MVA_5}{Mean vessel area chronology for tree 5}
  \item{MVA_6}{Mean vessel area chronology for tree 6}
  \item{MVA_7}{Mean vessel area chronology for tree 7}
  \item{MVA_8}{Mean vessel area chronology for tree 8}
  \item{MVA_9}{Mean vessel area chronology for tree 9}
  \item{MVA_10}{Mean vessel area chronology for tree 10}
}}
\source{
prof. dr. Tom Levanič, Slovenian Forestry Institute, Večna pot 2,
Ljubljana, Slovenia
}
\usage{
example_proxies_individual
}
\description{
A dataset of individual tree-ring chronologies from a lowland
forest in Slovenia. The first row represents a value of a year in 2015.
}
\keyword{datasets}
