% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter.R
\name{iter}
\alias{iter}
\title{iter}
\usage{
iter(formula, dataset, k, neurons, MT_M, MT_N, MT_U, MT_R, BMT_P, BMT_I, BMT_M,
  BMT_N, BMT_U, BMT_R, RF_P, RF_I, RF_depth, multiply)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model to be fitted.}

\item{dataset}{a data frame with dependent and independent variables as
columns and (optional) years as row names}

\item{k}{number of folds for cross-validation}

\item{neurons}{positive integer that indicates the number of neurons used
for brnn method}

\item{MT_M}{minimum number of instances used by model trees}

\item{MT_N}{unpruned (argument for model trees)}

\item{MT_U}{unsmoothed (argument for model trees)}

\item{MT_R}{use regression trees (argument for model trees)}

\item{BMT_P}{bagSizePercent (argument for bagging of model trees)}

\item{BMT_I}{number of iterations (argument for bagging of model trees)}

\item{BMT_M}{minimum number of instances used by model trees}

\item{BMT_N}{unpruned (argument for bagging of model trees)}

\item{BMT_U}{unsmoothed (argument for bagging of model trees)}

\item{BMT_R}{use regression trees (argument for bagging of model trees)}

\item{RF_P}{bagSizePercent (argument for random forest)}

\item{RF_I}{number of iterations (argument for random forest)}

\item{RF_depth}{maxDepth (argument for random forest)}

\item{multiply}{an intiger that will be used to change the seed options
for different repeats. set.seed(multiply*5)}
}
\value{
list with data frames of calculated measures for different methods
}
\description{
splits data into desired number of folds and performs cross-validation
with measure calculations
}
