% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNMI_daily_transform.R
\name{KNMI_daily_transform}
\alias{KNMI_daily_transform}
\title{KNMI_daily_transform}
\usage{
KNMI_daily_transform(input)
}
\arguments{
\item{input}{typical KNMI daily data format}
}
\value{
env_data suitable for daily_response()
}
\description{
transforms daily data obtained from KNMI Climate explorer into data frame suitable for daily_response()
}
\examples{
data(swit272_daily_temperatures)
proper_env_data <- KNMI_daily_transform(swit272_daily_temperatures)
}
