% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRE_daily_temperatures.R
\docType{data}
\name{KRE_daily_temperatures}
\alias{KRE_daily_temperatures}
\title{Daily mean temperatures for Ljubljana from 2017 - 1955}
\format{A data frame with 63 rows and 366 variables:
\describe{
  \item{X1}{Temperatures on the day 1 of a year}
  \item{X2}{Temperatures on the day 2 of a year}
  \item{X3}{Temperatures on the day 3 of a year}
  \item{X4}{Temperatures on the day 4 of a year}
  \item{X5}{Temperatures on the day 5 of a year}
  \item{X6}{Temperatures on the day 6 of a year}
  \item{X7}{Temperatures on the day 7 of a year}
  \item{X8}{Temperatures on the day 8 of a year}
  \item{X9}{Temperatures on the day 9 of a year}
  \item{X10}{Temperatures on the day 10 of a year}
  \item{X11}{Temperatures on the day 11 of a year}
  \item{X12}{Temperatures on the day 12 of a year}
  \item{X13}{Temperatures on the day 13 of a year}
  \item{X14}{Temperatures on the day 14 of a year}
  \item{X15}{Temperatures on the day 15 of a year}
  \item{X16}{Temperatures on the day 16 of a year}
  \item{X17}{Temperatures on the day 17 of a year}
  \item{X18}{Temperatures on the day 18 of a year}
  \item{X19}{Temperatures on the day 19 of a year}
  \item{X20}{Temperatures on the day 20 of a year}
  \item{X21}{Temperatures on the day 21 of a year}
  \item{X22}{Temperatures on the day 22 of a year}
  \item{X23}{Temperatures on the day 23 of a year}
  \item{X24}{Temperatures on the day 24 of a year}
  \item{X25}{Temperatures on the day 25 of a year}
  \item{X26}{Temperatures on the day 26 of a year}
  \item{X27}{Temperatures on the day 27 of a year}
  \item{X28}{Temperatures on the day 28 of a year}
  \item{X29}{Temperatures on the day 29 of a year}
  \item{X30}{Temperatures on the day 30 of a year}
  \item{X31}{Temperatures on the day 31 of a year}
  \item{X32}{Temperatures on the day 32 of a year}
  \item{X33}{Temperatures on the day 33 of a year}
  \item{X34}{Temperatures on the day 34 of a year}
  \item{X35}{Temperatures on the day 35 of a year}
  \item{X36}{Temperatures on the day 36 of a year}
  \item{X37}{Temperatures on the day 37 of a year}
  \item{X38}{Temperatures on the day 38 of a year}
  \item{X39}{Temperatures on the day 39 of a year}
  \item{X40}{Temperatures on the day 40 of a year}
  \item{X41}{Temperatures on the day 41 of a year}
  \item{X42}{Temperatures on the day 42 of a year}
  \item{X43}{Temperatures on the day 43 of a year}
  \item{X44}{Temperatures on the day 44 of a year}
  \item{X45}{Temperatures on the day 45 of a year}
  \item{X46}{Temperatures on the day 46 of a year}
  \item{X47}{Temperatures on the day 47 of a year}
  \item{X48}{Temperatures on the day 48 of a year}
  \item{X49}{Temperatures on the day 49 of a year}
  \item{X50}{Temperatures on the day 50 of a year}
  \item{X51}{Temperatures on the day 51 of a year}
  \item{X52}{Temperatures on the day 52 of a year}
  \item{X53}{Temperatures on the day 53 of a year}
  \item{X54}{Temperatures on the day 54 of a year}
  \item{X55}{Temperatures on the day 55 of a year}
  \item{X56}{Temperatures on the day 56 of a year}
  \item{X57}{Temperatures on the day 57 of a year}
  \item{X58}{Temperatures on the day 58 of a year}
  \item{X59}{Temperatures on the day 59 of a year}
  \item{X60}{Temperatures on the day 60 of a year}
  \item{X61}{Temperatures on the day 61 of a year}
  \item{X62}{Temperatures on the day 62 of a year}
  \item{X63}{Temperatures on the day 63 of a year}
  \item{X64}{Temperatures on the day 64 of a year}
  \item{X65}{Temperatures on the day 65 of a year}
  \item{X66}{Temperatures on the day 66 of a year}
  \item{X67}{Temperatures on the day 67 of a year}
  \item{X68}{Temperatures on the day 68 of a year}
  \item{X69}{Temperatures on the day 69 of a year}
  \item{X70}{Temperatures on the day 70 of a year}
  \item{X71}{Temperatures on the day 71 of a year}
  \item{X72}{Temperatures on the day 72 of a year}
  \item{X73}{Temperatures on the day 73 of a year}
  \item{X74}{Temperatures on the day 74 of a year}
  \item{X75}{Temperatures on the day 75 of a year}
  \item{X76}{Temperatures on the day 76 of a year}
  \item{X77}{Temperatures on the day 77 of a year}
  \item{X78}{Temperatures on the day 78 of a year}
  \item{X79}{Temperatures on the day 79 of a year}
  \item{X80}{Temperatures on the day 80 of a year}
  \item{X81}{Temperatures on the day 81 of a year}
  \item{X82}{Temperatures on the day 82 of a year}
  \item{X83}{Temperatures on the day 83 of a year}
  \item{X84}{Temperatures on the day 84 of a year}
  \item{X85}{Temperatures on the day 85 of a year}
  \item{X86}{Temperatures on the day 86 of a year}
  \item{X87}{Temperatures on the day 87 of a year}
  \item{X88}{Temperatures on the day 88 of a year}
  \item{X89}{Temperatures on the day 89 of a year}
  \item{X90}{Temperatures on the day 90 of a year}
  \item{X91}{Temperatures on the day 91 of a year}
  \item{X92}{Temperatures on the day 92 of a year}
  \item{X93}{Temperatures on the day 93 of a year}
  \item{X94}{Temperatures on the day 94 of a year}
  \item{X95}{Temperatures on the day 95 of a year}
  \item{X96}{Temperatures on the day 96 of a year}
  \item{X97}{Temperatures on the day 97 of a year}
  \item{X98}{Temperatures on the day 98 of a year}
  \item{X99}{Temperatures on the day 99 of a year}
  \item{X100}{Temperatures on the day 100 of a year}
  \item{X101}{Temperatures on the day 101 of a year}
  \item{X102}{Temperatures on the day 102 of a year}
  \item{X103}{Temperatures on the day 103 of a year}
  \item{X104}{Temperatures on the day 104 of a year}
  \item{X105}{Temperatures on the day 105 of a year}
  \item{X106}{Temperatures on the day 106 of a year}
  \item{X107}{Temperatures on the day 107 of a year}
  \item{X108}{Temperatures on the day 108 of a year}
  \item{X109}{Temperatures on the day 109 of a year}
  \item{X110}{Temperatures on the day 110 of a year}
  \item{X111}{Temperatures on the day 111 of a year}
  \item{X112}{Temperatures on the day 112 of a year}
  \item{X113}{Temperatures on the day 113 of a year}
  \item{X114}{Temperatures on the day 114 of a year}
  \item{X115}{Temperatures on the day 115 of a year}
  \item{X116}{Temperatures on the day 116 of a year}
  \item{X117}{Temperatures on the day 117 of a year}
  \item{X118}{Temperatures on the day 118 of a year}
  \item{X119}{Temperatures on the day 119 of a year}
  \item{X120}{Temperatures on the day 120 of a year}
  \item{X121}{Temperatures on the day 121 of a year}
  \item{X122}{Temperatures on the day 122 of a year}
  \item{X123}{Temperatures on the day 123 of a year}
  \item{X124}{Temperatures on the day 124 of a year}
  \item{X125}{Temperatures on the day 125 of a year}
  \item{X126}{Temperatures on the day 126 of a year}
  \item{X127}{Temperatures on the day 127 of a year}
  \item{X128}{Temperatures on the day 128 of a year}
  \item{X129}{Temperatures on the day 129 of a year}
  \item{X130}{Temperatures on the day 130 of a year}
  \item{X131}{Temperatures on the day 131 of a year}
  \item{X132}{Temperatures on the day 132 of a year}
  \item{X133}{Temperatures on the day 133 of a year}
  \item{X134}{Temperatures on the day 134 of a year}
  \item{X135}{Temperatures on the day 135 of a year}
  \item{X136}{Temperatures on the day 136 of a year}
  \item{X137}{Temperatures on the day 137 of a year}
  \item{X138}{Temperatures on the day 138 of a year}
  \item{X139}{Temperatures on the day 139 of a year}
  \item{X140}{Temperatures on the day 140 of a year}
  \item{X141}{Temperatures on the day 141 of a year}
  \item{X142}{Temperatures on the day 142 of a year}
  \item{X143}{Temperatures on the day 143 of a year}
  \item{X144}{Temperatures on the day 144 of a year}
  \item{X145}{Temperatures on the day 145 of a year}
  \item{X146}{Temperatures on the day 146 of a year}
  \item{X147}{Temperatures on the day 147 of a year}
  \item{X148}{Temperatures on the day 148 of a year}
  \item{X149}{Temperatures on the day 149 of a year}
  \item{X150}{Temperatures on the day 150 of a year}
  \item{X151}{Temperatures on the day 151 of a year}
  \item{X152}{Temperatures on the day 152 of a year}
  \item{X153}{Temperatures on the day 153 of a year}
  \item{X154}{Temperatures on the day 154 of a year}
  \item{X155}{Temperatures on the day 155 of a year}
  \item{X156}{Temperatures on the day 156 of a year}
  \item{X157}{Temperatures on the day 157 of a year}
  \item{X158}{Temperatures on the day 158 of a year}
  \item{X159}{Temperatures on the day 159 of a year}
  \item{X160}{Temperatures on the day 160 of a year}
  \item{X161}{Temperatures on the day 161 of a year}
  \item{X162}{Temperatures on the day 162 of a year}
  \item{X163}{Temperatures on the day 163 of a year}
  \item{X164}{Temperatures on the day 164 of a year}
  \item{X165}{Temperatures on the day 165 of a year}
  \item{X166}{Temperatures on the day 166 of a year}
  \item{X167}{Temperatures on the day 167 of a year}
  \item{X168}{Temperatures on the day 168 of a year}
  \item{X169}{Temperatures on the day 169 of a year}
  \item{X170}{Temperatures on the day 170 of a year}
  \item{X171}{Temperatures on the day 171 of a year}
  \item{X172}{Temperatures on the day 172 of a year}
  \item{X173}{Temperatures on the day 173 of a year}
  \item{X174}{Temperatures on the day 174 of a year}
  \item{X175}{Temperatures on the day 175 of a year}
  \item{X176}{Temperatures on the day 176 of a year}
  \item{X177}{Temperatures on the day 177 of a year}
  \item{X178}{Temperatures on the day 178 of a year}
  \item{X179}{Temperatures on the day 179 of a year}
  \item{X180}{Temperatures on the day 180 of a year}
  \item{X181}{Temperatures on the day 181 of a year}
  \item{X182}{Temperatures on the day 182 of a year}
  \item{X183}{Temperatures on the day 183 of a year}
  \item{X184}{Temperatures on the day 184 of a year}
  \item{X185}{Temperatures on the day 185 of a year}
  \item{X186}{Temperatures on the day 186 of a year}
  \item{X187}{Temperatures on the day 187 of a year}
  \item{X188}{Temperatures on the day 188 of a year}
  \item{X189}{Temperatures on the day 189 of a year}
  \item{X190}{Temperatures on the day 190 of a year}
  \item{X191}{Temperatures on the day 191 of a year}
  \item{X192}{Temperatures on the day 192 of a year}
  \item{X193}{Temperatures on the day 193 of a year}
  \item{X194}{Temperatures on the day 194 of a year}
  \item{X195}{Temperatures on the day 195 of a year}
  \item{X196}{Temperatures on the day 196 of a year}
  \item{X197}{Temperatures on the day 197 of a year}
  \item{X198}{Temperatures on the day 198 of a year}
  \item{X199}{Temperatures on the day 199 of a year}
  \item{X200}{Temperatures on the day 200 of a year}
  \item{X201}{Temperatures on the day 201 of a year}
  \item{X202}{Temperatures on the day 202 of a year}
  \item{X203}{Temperatures on the day 203 of a year}
  \item{X204}{Temperatures on the day 204 of a year}
  \item{X205}{Temperatures on the day 205 of a year}
  \item{X206}{Temperatures on the day 206 of a year}
  \item{X207}{Temperatures on the day 207 of a year}
  \item{X208}{Temperatures on the day 208 of a year}
  \item{X209}{Temperatures on the day 209 of a year}
  \item{X210}{Temperatures on the day 210 of a year}
  \item{X211}{Temperatures on the day 211 of a year}
  \item{X212}{Temperatures on the day 212 of a year}
  \item{X213}{Temperatures on the day 213 of a year}
  \item{X214}{Temperatures on the day 214 of a year}
  \item{X215}{Temperatures on the day 215 of a year}
  \item{X216}{Temperatures on the day 216 of a year}
  \item{X217}{Temperatures on the day 217 of a year}
  \item{X218}{Temperatures on the day 218 of a year}
  \item{X219}{Temperatures on the day 219 of a year}
  \item{X220}{Temperatures on the day 220 of a year}
  \item{X221}{Temperatures on the day 221 of a year}
  \item{X222}{Temperatures on the day 222 of a year}
  \item{X223}{Temperatures on the day 223 of a year}
  \item{X224}{Temperatures on the day 224 of a year}
  \item{X225}{Temperatures on the day 225 of a year}
  \item{X226}{Temperatures on the day 226 of a year}
  \item{X227}{Temperatures on the day 227 of a year}
  \item{X228}{Temperatures on the day 228 of a year}
  \item{X229}{Temperatures on the day 229 of a year}
  \item{X230}{Temperatures on the day 230 of a year}
  \item{X231}{Temperatures on the day 231 of a year}
  \item{X232}{Temperatures on the day 232 of a year}
  \item{X233}{Temperatures on the day 233 of a year}
  \item{X234}{Temperatures on the day 234 of a year}
  \item{X235}{Temperatures on the day 235 of a year}
  \item{X236}{Temperatures on the day 236 of a year}
  \item{X237}{Temperatures on the day 237 of a year}
  \item{X238}{Temperatures on the day 238 of a year}
  \item{X239}{Temperatures on the day 239 of a year}
  \item{X240}{Temperatures on the day 240 of a year}
  \item{X241}{Temperatures on the day 241 of a year}
  \item{X242}{Temperatures on the day 242 of a year}
  \item{X243}{Temperatures on the day 243 of a year}
  \item{X244}{Temperatures on the day 244 of a year}
  \item{X245}{Temperatures on the day 245 of a year}
  \item{X246}{Temperatures on the day 246 of a year}
  \item{X247}{Temperatures on the day 247 of a year}
  \item{X248}{Temperatures on the day 248 of a year}
  \item{X249}{Temperatures on the day 249 of a year}
  \item{X250}{Temperatures on the day 250 of a year}
  \item{X251}{Temperatures on the day 251 of a year}
  \item{X252}{Temperatures on the day 252 of a year}
  \item{X253}{Temperatures on the day 253 of a year}
  \item{X254}{Temperatures on the day 254 of a year}
  \item{X255}{Temperatures on the day 255 of a year}
  \item{X256}{Temperatures on the day 256 of a year}
  \item{X257}{Temperatures on the day 257 of a year}
  \item{X258}{Temperatures on the day 258 of a year}
  \item{X259}{Temperatures on the day 259 of a year}
  \item{X260}{Temperatures on the day 260 of a year}
  \item{X261}{Temperatures on the day 261 of a year}
  \item{X262}{Temperatures on the day 262 of a year}
  \item{X263}{Temperatures on the day 263 of a year}
  \item{X264}{Temperatures on the day 264 of a year}
  \item{X265}{Temperatures on the day 265 of a year}
  \item{X266}{Temperatures on the day 266 of a year}
  \item{X267}{Temperatures on the day 267 of a year}
  \item{X268}{Temperatures on the day 268 of a year}
  \item{X269}{Temperatures on the day 269 of a year}
  \item{X270}{Temperatures on the day 270 of a year}
  \item{X271}{Temperatures on the day 271 of a year}
  \item{X272}{Temperatures on the day 272 of a year}
  \item{X273}{Temperatures on the day 273 of a year}
  \item{X274}{Temperatures on the day 274 of a year}
  \item{X275}{Temperatures on the day 275 of a year}
  \item{X276}{Temperatures on the day 276 of a year}
  \item{X277}{Temperatures on the day 277 of a year}
  \item{X278}{Temperatures on the day 278 of a year}
  \item{X279}{Temperatures on the day 279 of a year}
  \item{X280}{Temperatures on the day 280 of a year}
  \item{X281}{Temperatures on the day 281 of a year}
  \item{X282}{Temperatures on the day 282 of a year}
  \item{X283}{Temperatures on the day 283 of a year}
  \item{X284}{Temperatures on the day 284 of a year}
  \item{X285}{Temperatures on the day 285 of a year}
  \item{X286}{Temperatures on the day 286 of a year}
  \item{X287}{Temperatures on the day 287 of a year}
  \item{X288}{Temperatures on the day 288 of a year}
  \item{X289}{Temperatures on the day 289 of a year}
  \item{X290}{Temperatures on the day 290 of a year}
  \item{X291}{Temperatures on the day 291 of a year}
  \item{X292}{Temperatures on the day 292 of a year}
  \item{X293}{Temperatures on the day 293 of a year}
  \item{X294}{Temperatures on the day 294 of a year}
  \item{X295}{Temperatures on the day 295 of a year}
  \item{X296}{Temperatures on the day 296 of a year}
  \item{X297}{Temperatures on the day 297 of a year}
  \item{X298}{Temperatures on the day 298 of a year}
  \item{X299}{Temperatures on the day 299 of a year}
  \item{X300}{Temperatures on the day 300 of a year}
  \item{X301}{Temperatures on the day 301 of a year}
  \item{X302}{Temperatures on the day 302 of a year}
  \item{X303}{Temperatures on the day 303 of a year}
  \item{X304}{Temperatures on the day 304 of a year}
  \item{X305}{Temperatures on the day 305 of a year}
  \item{X306}{Temperatures on the day 306 of a year}
  \item{X307}{Temperatures on the day 307 of a year}
  \item{X308}{Temperatures on the day 308 of a year}
  \item{X309}{Temperatures on the day 309 of a year}
  \item{X310}{Temperatures on the day 310 of a year}
  \item{X311}{Temperatures on the day 311 of a year}
  \item{X312}{Temperatures on the day 312 of a year}
  \item{X313}{Temperatures on the day 313 of a year}
  \item{X314}{Temperatures on the day 314 of a year}
  \item{X315}{Temperatures on the day 315 of a year}
  \item{X316}{Temperatures on the day 316 of a year}
  \item{X317}{Temperatures on the day 317 of a year}
  \item{X318}{Temperatures on the day 318 of a year}
  \item{X319}{Temperatures on the day 319 of a year}
  \item{X320}{Temperatures on the day 320 of a year}
  \item{X321}{Temperatures on the day 321 of a year}
  \item{X322}{Temperatures on the day 322 of a year}
  \item{X323}{Temperatures on the day 323 of a year}
  \item{X324}{Temperatures on the day 324 of a year}
  \item{X325}{Temperatures on the day 325 of a year}
  \item{X326}{Temperatures on the day 326 of a year}
  \item{X327}{Temperatures on the day 327 of a year}
  \item{X328}{Temperatures on the day 328 of a year}
  \item{X329}{Temperatures on the day 329 of a year}
  \item{X330}{Temperatures on the day 330 of a year}
  \item{X331}{Temperatures on the day 331 of a year}
  \item{X332}{Temperatures on the day 332 of a year}
  \item{X333}{Temperatures on the day 333 of a year}
  \item{X334}{Temperatures on the day 334 of a year}
  \item{X335}{Temperatures on the day 335 of a year}
  \item{X336}{Temperatures on the day 336 of a year}
  \item{X337}{Temperatures on the day 337 of a year}
  \item{X338}{Temperatures on the day 338 of a year}
  \item{X339}{Temperatures on the day 339 of a year}
  \item{X340}{Temperatures on the day 340 of a year}
  \item{X341}{Temperatures on the day 341 of a year}
  \item{X342}{Temperatures on the day 342 of a year}
  \item{X343}{Temperatures on the day 343 of a year}
  \item{X344}{Temperatures on the day 344 of a year}
  \item{X345}{Temperatures on the day 345 of a year}
  \item{X346}{Temperatures on the day 346 of a year}
  \item{X347}{Temperatures on the day 347 of a year}
  \item{X348}{Temperatures on the day 348 of a year}
  \item{X349}{Temperatures on the day 349 of a year}
  \item{X350}{Temperatures on the day 350 of a year}
  \item{X351}{Temperatures on the day 351 of a year}
  \item{X352}{Temperatures on the day 352 of a year}
  \item{X353}{Temperatures on the day 353 of a year}
  \item{X354}{Temperatures on the day 354 of a year}
  \item{X355}{Temperatures on the day 355 of a year}
  \item{X356}{Temperatures on the day 356 of a year}
  \item{X357}{Temperatures on the day 357 of a year}
  \item{X358}{Temperatures on the day 358 of a year}
  \item{X359}{Temperatures on the day 359 of a year}
  \item{X360}{Temperatures on the day 360 of a year}
  \item{X361}{Temperatures on the day 361 of a year}
  \item{X362}{Temperatures on the day 362 of a year}
  \item{X363}{Temperatures on the day 363 of a year}
  \item{X364}{Temperatures on the day 364 of a year}
  \item{X365}{Temperatures on the day 365 of a year}
  \item{X366}{Temperatures on the day 366 of a year}
}}
\source{
\url{http://meteo.arso.gov.si/met/sl/archive/}
}
\usage{
KRE_daily_temperatures
}
\description{
A dataset of daily mean temperatures in Kredarica (Slovenia). The first row
represents temperatures in 1955. The first column represents the first day
of a year, the second column represents the second day of a year, etc.
Row names represent years.
}
\keyword{datasets}
