% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LJ_monthly_precipitation.R
\docType{data}
\name{LJ_monthly_precipitation}
\alias{LJ_monthly_precipitation}
\title{Monthly sums of precipitation for Ljubljana from 2018 - 1900. Tidy format.}
\format{
A data frame with 1417 rows and 3 variables:
\describe{
  \item{Year}{year}
  \item{Month}{Month}
  \item{Precipitation}{Sum of precipitation}
}
}
\source{
\url{http://climexp.knmi.nl/start.cgi}
}
\usage{
LJ_monthly_precipitation
}
\description{
A dataset of monthly sums of precipitations in Ljubljana (Slovenia). The first row
represents precipitation sum for January 1900.
}
\keyword{datasets}
