% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LJ_monthly_temperatures.R
\docType{data}
\name{LJ_monthly_temperatures}
\alias{LJ_monthly_temperatures}
\title{Monthly mean air temperatures for Ljubljana from 2015 - 1900}
\format{
A data frame with 116 rows and 12 variables:
\describe{
  \item{Jan}{Mean monthly air temperature for January from 1900 to 2015}
  \item{Feb}{Mean monthly air temperature for February from 1900 to 2015}
  \item{Mar}{Mean monthly air temperature for March from 1900 to 2015}
  \item{Apr}{Mean monthly air temperature for April from 1900 to 2015}
  \item{May}{Mean monthly air temperature for May from 1900 to 2015}
  \item{Jun}{Mean monthly air temperature for June from 1900 to 2015}
  \item{Jul}{Mean monthly air temperature for July from 1900 to 2015}
  \item{Aug}{Mean monthly air temperature for August from 1900 to 2015}
  \item{Sep}{Mean monthly air temperature for September from 1900 to 2015}
  \item{Oct}{Mean monthly air temperature for October from 1900 to 2015}
  \item{Nov}{Mean monthly air temperature for November from 1900 to 2015}
  \item{Dec}{Mean monthly air temperature for December from 1900 to 2015}
}
}
\source{
\url{http://meteo.arso.gov.si/met/sl/archive/}
}
\usage{
LJ_monthly_temperatures
}
\description{
A dataset of monthly mean air temperatures in Ljubljana (Slovenia). The
first row represents temperatures in 2015. The first column represents
mean January temperature, the second column represents mean February
temperature. etc. Row names represent year.
}
\keyword{datasets}
