% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extreme.R
\name{plot_extreme}
\alias{plot_extreme}
\title{plot_extreme}
\usage{
plot_extreme(
  result_daily_response,
  title = TRUE,
  ylimits = NULL,
  reference_window = "start",
  type = "daily",
  show_year_separators = TRUE
)
}
\arguments{
\item{result_daily_response}{a list with objects as produced by
daily_response or monthly_response function}

\item{title}{logical, if set to FALSE, no plot title is displayed}

\item{ylimits}{limit of the y axes. It should be given as ylimits = c(0,1)}

\item{reference_window}{character string, the reference_window argument describes,
how each calculation is referred. There are three different options: 'start'
(default), 'end' and 'middle'. If available, the value stored in
result_daily_response$reference_window is used.}

\item{type}{the character string describing type of analysis: daily or monthly}

\item{show_year_separators}{logical. If TRUE, dashed vertical lines are
added between relative-year blocks. The default is TRUE.}
}
\value{
A ggplot2 object containing the plot display
}
\description{
Graphs a line or bar plot of a row with the highest metric in a matrix,
produced by \code{\link{daily_response}} or \code{\link{monthly_response}}
functions. Bar plot is drawn for monthly_response(), while for daily_response,
line plot is produced.
}
\keyword{internal}
