% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specific.R
\name{plot_specific}
\alias{plot_specific}
\title{plot_specific}
\usage{
plot_specific(result_daily_response, window_width, title = TRUE,
  ylimits = NULL)
}
\arguments{
\item{result_daily_response}{a list with three objects as produced by
daily_response function}

\item{window_width}{integer representing window width to be displayed}

\item{title}{logical, if set to FALSE, no plot title is displayed}

\item{ylimits}{limit of the y axes. It should be given as ylimits = c(0,1)}
}
\value{
A ggplot2 object containing the plot display
}
\description{
Graphs a line plot of a row with a selected window width in a matrix,
produced by \code{\link{daily_response}} function.
}
\keyword{internal}
