% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{round2}
\usage{
round2(x, n = 0)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{integer indicating the number of decimal places}
}
\value{
a numeric vector with rounded numbers
}
\description{
Rounding function, where numbers are rounded up, 0.5 = 1
}
\references{
http://alandgraf.blogspot.com/2012/06/rounding-in-r.html
}
\keyword{internal}
