% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swit272.R
\docType{data}
\name{swit272}
\alias{swit272}
\title{Standardised tree-ring width chronology swit272, Larix decidua Mill.}
\format{A data frame with 273 rows and 1 variable:
\describe{
  \item{TRWi}{Standardised TRW index}
}}
\source{
\url{https://www.ncdc.noaa.gov/paleo/study/14108}
}
\usage{
swit272
}
\description{
A TRW chronology swit272
Investigators: Bigler, C.; Claluna, A.
Site_Name:  Sils-Maria GR Blais dal Fo
Location: Switzerland
Northernmost_Latitude: 46.4333
Southernmost_Latitude: 46.4333
Easternmost_Longitude: 9.7833
Westernmost_Longitude: 9.7833
Elevation: 2100
}
\keyword{datasets}
