% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swit272_daily_temperatures.R
\docType{data}
\name{swit272_daily_temperatures}
\alias{swit272_daily_temperatures}
\title{Daily temperatures for swit272_daily_temperatures}
\format{A data frame with 43068 rows and 3 variables:
\describe{
  \item{V1}{character string describing date}
  \item{V2}{mean temperature}
}}
\source{
\url{https://climexp.knmi.nl/}
}
\usage{
swit272_daily_temperatures
}
\description{
Daily mean temperature (Celsius) downloaded from KNMI Climate explorer, daily fields: E-OBS 1950-now 0.25°.
This is gridded data for grid point: lon = 9.875, lat = 46.375.
}
\keyword{datasets}
