% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ones.R
\name{count_ones}
\alias{count_ones}
\title{count_ones}
\usage{
count_ones(vector)
}
\arguments{
\item{vector}{a vector of integers}
}
\value{
an integer of counted ones
}
\description{
calculates share of integer 1 in a vector
}
\examples{
vector_1 <- seq(1:10)
# count_ones(vector_1)
}
\keyword{internal}
