% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_response_seascorr.R
\name{daily_response_seascorr}
\alias{daily_response_seascorr}
\title{daily_response_seascorr}
\usage{
daily_response_seascorr(
  response,
  env_data_primary,
  env_data_control,
  lower_limit = 30,
  upper_limit = 90,
  fixed_width = 0,
  previous_year = FALSE,
  pcor_method = "pearson",
  remove_insignificant = TRUE,
  alpha = 0.05,
  row_names_subset = FALSE,
  PCA_transformation = FALSE,
  log_preprocess = TRUE,
  components_selection = "automatic",
  eigenvalues_threshold = 1,
  N_components = 2,
  aggregate_function_env_data_primary = "mean",
  aggregate_function_env_data_control = "mean",
  temporal_stability_check = "sequential",
  k = 2,
  k_running_window = 30,
  subset_years = NULL,
  plot_specific_window = NULL,
  ylimits = NULL,
  seed = NULL,
  tidy_env_data_primary = FALSE,
  tidy_env_data_control = FALSE,
  reference_window = "start",
  boot = FALSE,
  boot_n = 1000,
  boot_ci_type = "norm",
  boot_conf_int = 0.95,
  day_interval = ifelse(c(previous_year == TRUE, previous_year == TRUE), c(-1, 366),
    c(1, 366)),
  dc_method = NULL,
  dc_nyrs = NULL,
  dc_f = 0.5,
  dc_pos.slope = FALSE,
  dc_constrain.nls = c("never", "when.fail", "always"),
  dc_span = "cv",
  dc_bass = 0,
  dc_difference = FALSE,
  pcor_na_use = "pairwise.complete",
  skip_window_length = 1,
  skip_window_position = 1
)
}
\arguments{
\item{response}{a data frame with tree-ring proxy variable and (optional)
years as row names. Row.names should be matched with those from env_data_primary
and env_data_control data frame. If not, set the row_names_subset argument to
TRUE.}

\item{env_data_primary}{primary data frame of daily sequences of environmental
data as columns and years as row names. Each row represents a year and
each column represents a day of a year. Row.names should be matched with
those from the response data frame. If not, set the argument row_names_subset
to TRUE. Alternatively, env_data_primary could be a tidy data with three columns,
i.e. Year, DOY and third column representing values of mean temperatures,
sum of precipitation etc. If tidy data is passed to the function, set the argument
tidy_env_data_primary to TRUE.}

\item{env_data_control}{a data frame of daily sequences of environmental data as
columns and years as row names. This data is used as control for calculations of
partial correlation coefficients. Each row represents a year and each column
represents a day of a year. Row.names should be matched with those from the
response data frame. If not, set the row_names_subset argument to TRUE.
Alternatively, env_data_control could be a tidy data with three columns,
i.e. Year, DOY and third column representing values of mean temperatures, sum
of precipitation etc. If tidy data is passed to the function, set the argument
tidy_env_data_control to TRUE.}

\item{lower_limit}{lower limit of window width}

\item{upper_limit}{upper limit of window width}

\item{fixed_width}{fixed width used for calculation. If fixed_width is
assigned a value, upper_limit and lower_limit will be ignored}

\item{previous_year}{if set to TRUE, env_data_primary, env_data_control  and
response variables will be rearranged in a way, that also previous year will
be used for calculations of selected statistical metric.}

\item{pcor_method}{a character string indicating which partial correlation
coefficient is to be computed. One of "pearson" (default), "kendall", or
"spearman", can be abbreviated.}

\item{remove_insignificant}{if set to TRUE, removes all correlations bellow
the significant threshold level, based on a selected alpha.}

\item{alpha}{significance level used to remove insignificant calculations.}

\item{row_names_subset}{if set to TRUE, row.names are used to subset
env_data_primary, env_data_control and response data frames. Only years from
all three data frames are kept.}

\item{PCA_transformation}{if set to TRUE, all variables in the response
data frame will be transformed using PCA transformation.}

\item{log_preprocess}{if set to TRUE, variables will be transformed with
logarithmic transformation before used in PCA}

\item{components_selection}{character string specifying how to select the Principal
Components used as predictors.
There are three options: "automatic", "manual" and "plot_selection". If
argument is set to automatic, all scores with eigenvalues above 1 will be
selected. This threshold could be changed by changing the
eigenvalues_threshold argument. If parameter is set to "manual", user should
set the number of components with N_components argument. If components
selection is set to "plot_selection", Scree plot will be shown and a user must
manually enter the number of components to be used as predictors.}

\item{eigenvalues_threshold}{threshold for automatic selection of Principal Components}

\item{N_components}{number of Principal Components used as predictors}

\item{aggregate_function_env_data_primary}{character string specifying how the
daily data from env_data_primary should be aggregated. The default is 'mean',
the other options are 'median', 'sum', 'min' and 'max'}

\item{aggregate_function_env_data_control}{character string specifying how the
daily data from env_data_control should be aggregated. The default is 'mean',
the other options are 'median', 'sum', 'min' and 'max'}

\item{temporal_stability_check}{character string, specifying, how temporal stability
between the optimal selection and response variable(s) will be analysed. Current
possibilities are "sequential", "progressive" and "running_window". Sequential check
will split data into k splits and calculate selected metric for each split. Progressive
check will split data into k splits, calculate metric for the first split and then
progressively add 1 split at a time and calculate selected metric. For running window,
select the length of running window with the k_running_window argument.}

\item{k}{integer, number of breaks (splits) for temporal stability}

\item{k_running_window}{the length of running window for temporal stability check.
Applicable only if temporal_stability argument is set to running window.}

\item{subset_years}{a subset of years to be analyzed. Should be given in the form of
subset_years = c(1980, 2005)}

\item{plot_specific_window}{integer representing window width to be displayed
for plot_specific}

\item{ylimits}{limit of the y axes for plot_extreme and plot_specific. It should be
given in the form of: ylimits = c(0,1)}

\item{seed}{optional seed argument for reproducible results}

\item{tidy_env_data_primary}{if set to TRUE, env_data_primary should be inserted as a
data frame with three columns: "Year", "DOY", "Precipitation/Temperature/etc."}

\item{tidy_env_data_control}{if set to TRUE, env_data_control should be inserted as a
data frame with three columns: "Year", "DOY", "Precipitation/Temperature/etc."}

\item{reference_window}{character string, the reference_window argument describes,
how each calculation is referred. There are three different options: 'start'
(default), 'end' and 'middle'. If the reference_window argument is set to 'start',
then each calculation is related to the starting day of window. If the
reference_window argument is set to 'middle', each calculation is related to the
middle day of window calculation. If the reference_window argument is set to
'end', then each calculation is related to the ending day of window calculation.
For example, if we consider correlations with window from DOY 15 to DOY 35. If
reference window is set to  'start', then this calculation will be related to the
DOY 15. If the reference window is set to 'end', then this calculation will be
related to the DOY 35. If the reference_window is set to 'middle', then this
calculation is related to DOY 25.
The optimal selection, which describes the optimal consecutive days that returns
the highest calculated metric and is obtained by the $plot_extreme output, is the
same for all three reference windows.}

\item{boot}{logical, if TRUE, bootstrap procedure will be used to calculate
partial correlation coefficients}

\item{boot_n}{The number of bootstrap replicates}

\item{boot_ci_type}{A character string representing the type of bootstrap intervals
required. The value should be any subset of the values c("norm","basic", "stud",
"perc", "bca").}

\item{boot_conf_int}{A scalar or vector containing the confidence level(s) of
the required interval(s)}

\item{day_interval}{a vector of two values: lower and upper time interval of
days that will be used to calculate statistical metrics. Negative values
indicate previous growing season days. This argument overwrites the calculation
limits defined by lower_limit and upper_limit arguments.}

\item{dc_method}{a character string to determine the method to detrend climate
(environmental) data.  Possible values are "none" (default), "SLD","Spline",
"ModNegExp", "Mean", "Friedman", "ModHugershoff". "SLD" refers to Simple
Linear Detrending}

\item{dc_nyrs}{a number giving the rigidity of the smoothing spline, defaults
to 0.67 of series length if nyrs is NULL (see dplR R package)}

\item{dc_f}{a number between 0 and 1 giving the frequency response or wavelength
cutoff. Defaults to 0.5 (see dplR R package)}

\item{dc_pos.slope}{a logical flag. Will allow for a positive slope to be used
in method "ModNegExp" and "ModHugershoff". If FALSE the line will be horizontal
(see dplR R package)}

\item{dc_constrain.nls}{a character string which controls the constraints of
the "ModNegExp" model and the "ModHugershoff"  (see dplR R package).}

\item{dc_span}{a numeric value controlling method "Friedman", or "cv" (default)
for automatic choice by cross-validation (see dplR R package)}

\item{dc_bass}{a numeric value controlling the smoothness of the fitted curve
in method "Friedman" (see dplR R package)}

\item{dc_difference}{a logical flag. Compute residuals by subtraction if TRUE,
otherwise use division (see dplR R package)}

\item{pcor_na_use}{an optional character string giving a method for computing
covariances in the presence of missing values for partial correlation
coefficients. This must be (an abbreviation of) one of the strings "all.obs",
"everything", "complete.obs", "na.or.complete", or "pairwise.complete.obs"
(default). See also the documentation for the base partial.r in psych R package}

\item{skip_window_length}{an integer specifying the frequency of window
selection for the calculations of climate-growth relationships. The default
value is 1, indicating that every window is included in the calculations.
When set to a value greater than 1, the function selectively processes
windows at regular intervals defined by this parameter. For instance, if
skip_window_length = 2, the function processes every second window.
Similarly, if skip_window_length = 3, every third window is processed,
skipping two windows in between each selected one. This parameter allows for
controlling the granularity of the analysis and can help in reducing
computation time by focusing on a subset of the data.}

\item{skip_window_position}{an integer specifying the frequency of window
positions used in the calculations of climate-growth relationships. The
default value is 1, indicating that every window position is included in the
calculations. When set to a value greater than 1, the function selectively
processes window positions at regular intervals defined by this parameter.
For instance, if skip_window_position = 2, the function processes every
second window position. Similarly, if skip_window_position = 3, every third
window position is processed, skipping two positions in between each selected
one. This parameter allows for controlling the granularity of the analysis
and can help in reducing computation time by focusing on a subset of the data.}
}
\value{
a list with 15 elements:
\enumerate{
 \item $calculations - a matrix with calculated metrics
 \item $method - the character string of a method
 \item $metric - the character string indicating the metric used for calculations
 \item $analysed_period - the character string specifying the analysed period based on the information from row names. If there are no row names, this argument is given as NA
 \item $optimized_return - data frame with two columns, response variable and aggregated (averaged) daily data that return the optimal results. This data.frame could be directly used to calibrate a model for climate reconstruction
 \item $optimized_return_all - a data frame with aggregated daily data, that returned the optimal result for the entire env_data_primary (and not only subset of analysed years)
 \item $transfer_function - a ggplot object: scatter plot of optimized return and a transfer line of the selected method
 \item $temporal_stability - a data frame with calculations of selected metric for different temporal subsets
 \item $cross_validation - not available for partial correlations
 \item $plot_heatmap - ggplot2 object: a heatmap of calculated metrics
 \item $plot_extreme - ggplot2 object: line plot of a row with the highest value in a matrix of calculated metrics
 \item $plot_specific -  ggplot2 object: line plot of a row with a selected window width in a matrix of calculated metrics
 \item $PCA_output - princomp object: the result output of the PCA analysis
 \item $type - the character string describing type of analysis: daily or monthly
 \item $reference_window - character string, which reference window was used for calculations
 \item $aggregated_climate_primary - matrix with all aggregated climate series of primary data
 \item $aggregated_climate_control - matrix with all aggregated climate series of control data
}
}
\description{
Function calculates all possible partial correlation coefficients between
tree-ring chronology and daily environmental (usually climate) data.
Calculations are based on moving window which is defined with two
arguments: lower_limit and upper_limit. All calculated (partial) correlation
coefficients are stored in a matrix. The location of stored correlation
in the matrix is indicating a window width (row names) and a location in a
matrix of daily sequences of environmental data (column names).
}
\examples{
\donttest{
# Load the dendroTools R package
library(dendroTools)

# Load data
data(data_MVA)
data(data_TRW)
data(data_TRW_1)
data(example_proxies_individual)
data(example_proxies_1)
data(LJ_daily_temperatures)
data(LJ_daily_precipitation)

# 1 Basic example
example_basic <- daily_response_seascorr(response = data_MVA,
                          env_data_primary = LJ_daily_temperatures,
                          env_data_control = LJ_daily_precipitation,
                          row_names_subset = TRUE,
                          lower_limit = 35, upper_limit = 45,
                          remove_insignificant = FALSE,
                          aggregate_function_env_data_primary = 'median',
                          aggregate_function_env_data_control = 'median',
                          alpha = 0.05, pcor_method = "spearman",
                          tidy_env_data_primary = FALSE,
                          previous_year = FALSE, boot = TRUE,
                          tidy_env_data_control = TRUE, boot_n = 10,
                          reference_window = "end", k = 5,
                          dc_method = "SLD",
                          day_interval = c(-100, 250),
                          skip_window_length = 1,
                          skip_window_position = 1
                          )
summary(example_basic)
plot(example_basic, type = 1)
plot(example_basic, type = 2)
plot(example_basic, type = 3)
example_basic$optimized_return
example_basic$optimized_return_all
example_basic$temporal_stability

# 2 Example with fixed temporal time window
example_fixed_width <- daily_response_seascorr(response = data_MVA,
                          env_data_primary = LJ_daily_temperatures,
                          env_data_control = LJ_daily_precipitation,
                          row_names_subset = TRUE,
                          remove_insignificant = TRUE,
                          aggregate_function_env_data_primary = 'mean',
                          aggregate_function_env_data_control = 'mean',
                          alpha = 0.05,
                          dc_method = "SLD",
                          fixed_width = 45,
                          tidy_env_data_primary = FALSE,
                          tidy_env_data_control = TRUE,
                          reference_window = "end")

summary(example_fixed_width)
plot(example_fixed_width, type = 1)
plot(example_fixed_width, type = 2)
example_fixed_width$optimized_return
example_fixed_width$optimized_return_all

}
}
