% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LJ_daily_precipitation.R
\docType{data}
\name{LJ_daily_precipitation}
\alias{LJ_daily_precipitation}
\title{Daily precipitation for Ljubljana from 2017 - 1899}
\format{A data frame with 43068 rows and 3 variables:
\describe{
  \item{Year}{year}
  \item{DOY}{day of year}
  \item{Precipitation}{Sum of precipitation}
}}
\source{
\url{http://climexp.knmi.nl/start.cgi}
}
\usage{
LJ_daily_precipitation
}
\description{
A dataset of daily sum of precipitation in Ljubljana (Slovenia). The first row
represents precipitation in 1899.
}
\keyword{datasets}
