% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(result_daily_response, reference_window = "start")
}
\arguments{
\item{result_daily_response}{a list with three objects as produced by
\code{\link{daily_response}} function}

\item{reference_window}{character string, the reference_window argument describes,
how each calculation is referred. There are three different options: 'start'
(default), 'end' and 'middle'. If the reference_window argument is set to 'start',
then each calculation is related to the starting day of window. If the
reference_window argument is set to 'middle', each calculation is related to the
middle day of window calculation. If the reference_window argument is set to
'end', then each calculation is related to the ending day of window calculation.
For example, if we consider correlations with window from DOY 15 to DOY 35. If
reference window is set to ‘start’, then this calculation will be related to the
DOY 15. If the reference window is set to ‘end’, then this calculation will be
related to the DOY 35. If the reference_window is set to 'middle', then this
calculation is related to DOY 25.}
}
\value{
A ggplot2 object containing the heatmap display
}
\description{
Graphs a heatmap of values stored in a matrix, such as produced
by \code{\link{daily_response}} function.
}
\examples{
\dontrun{
data(LJ_daily_temperatures)
data(example_proxies_1)
Example1 <- daily_response(response = example_proxies_1,
env_data = LJ_daily_temperatures, method = "lm", metric = "r.squared",
fixed_width = 90, previous_year = TRUE, row_names_subset = TRUE)
plot_heatmap(Example1)

Example2 <- daily_response(response = example_proxies_1,
env_data = LJ_daily_temperatures, method = "brnn",
metric = "adj.r.squared", lower_limit = 50, upper_limit = 55,
row_names_subset = TRUE)
plot_heatmap(Example2)

library(dplyr)
oxygen_isotope <- dplyr::select(example_proxies_1, O18)
Example3 <- daily_response(response = oxygen_isotope,
env_data = LJ_daily_temperatures, method = "cor", lower_limit = 50,
upper_limit = 55, previous_year = TRUE, row_names_subset = TRUE)
plot_heatmap(Example3)
}

}
\keyword{internal}
