% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swit272_daily_temperatures.R
\docType{data}
\name{swit272_daily_temperatures}
\alias{swit272_daily_temperatures}
\title{Daily temperatures for swit272_daily_temperatures}
\format{A data frame with 43068 rows and 3 variables:
\describe{
  \item{V1}{character string describing date}
  \item{V2}{mean temperature}
}}
\source{
\url{https://climexp.knmi.nl/}
}
\usage{
swit272_daily_temperatures
}
\description{
Daily mean temperature (Celsius) downloaded from KNMI Climate explorer, daily fields: E-OBS 1950-now 0.25°.
This is gridded data for grid point: lon = 9.875, lat = 46.375.
}
\details{
We acknowledge the E-OBS dataset from the EU-FP6 project UERRA (http://www.uerra.eu) and the Copernicus Climate Change Service, and the data providers in the ECA&D project (https://www.ecad.eu). Cornes, R., G. van der Schrier, E.J.M. van den Besselaar, and P.D. Jones. 2018: An Ensemble Version of the E-OBS Temperature and Precipitation Datasets, J. Geophys. Res. Atmos., 123. doi:10.1029/2017JD028200
}
\keyword{datasets}
